requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
asteroid-filterbanks==0.4.0
attrs==24.2.0
audioread==3.0.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
contourpy==1.3.0
coverage==5.5
cycler==0.12.1
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
einops==0.8.0
eyed3==0.9.7
ffmpeg-python==0.2.0
filelock==3.16.1
filetype==1.2.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
hmmlearn==0.3.2
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
HyperPyYAML==1.2.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
imbalanced-learn==0.12.3
imblearn==0.0
inquirerpy==0.3.4
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.7
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.8.4
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.50.2
openai-whisper @ git+https://github.com/openai/whisper.git@423492dda7806206abe56bdfe427c1096473a020
optuna==4.0.0
packaging==24.1
pandas==2.2.3
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
primePy==1.3
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==5.28.2
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==17.0.0
pyAudioAnalysis==0.3.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.1
SQLAlchemy==2.0.35
streamlit==1.38.0
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.7.0
toml==0.10.2
torch==2.4.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.4.1
torchmetrics==1.4.2
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==4.0.2
wcwidth==0.2.13
whisper==1.1.10
yarl==1.13.1
